requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.5
django-constance==4.2.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
whitenoise==6.8.2
